refactor(invitation): move InvitationQRCode component to features/invitation and update imports
