+++ features/invitation/components/invitation-qr-code.tsx
